requirements.txt
Django==2.2.1
django-pagedown==1.0.6
djangorestframework==3.9.4
Markdown==3.1.1
Pillow==6.0.0
psycopg2==2.8.2
python-slugify==3.0.2
pytz==2019.1
sqlparse==0.3.0
text-unidecode==1.2
